* refactor: ResponseEntity 응답 간략화

* feat: 노선별 추가 요금 적용

* docs: 기능 요구사항 업데이트

* feat: 연령별 요금 할인

* refactor: PathService 생성

* refactor: 요금 계산 Fare 클래스 생성

* test: @MockMvc 컨트롤러 테스트 추가

* refactor: 불필요한 Age 필드 제거

* refactor: station 검색 반복문 제거

* refactor: DijkstraShortestPath 생성과 WeightedMultigraph 초기화 하나의 메서드에서 처리

* refactor: @RequestParam null 예외 처리

* chore: validation 의존성 추가

* refactor: @Valid 파라미터 검증

* refactor: 경로 조회 예외 추가

* refactor: 누락된 @Override, @Transactional 애노테이션 적용

* refactor: Path 생성 인터페이스로 분리

* test: Path 객체 생성 확인

* test: 인수 테스트 시나리오 작성 및 테스트 보강

* refactor: 불필요한 import 제거

* style: dependency 선언 순서 변경

* docs: 연령별 요금 정책 추가

* feat: 연령별 요금 정책 추가

* refactor: 변수명 Ids -> stationIds로 변경

* style: 줄바꿈 컨벤션

* refactor: 메서드 네임 명확하게 변경

* refactor: 요금 계산 한번에 처리

* refactor: PathResponse 객체를 검증하도록 변경

* refactor: age 정렬 추가

* refactor: Path 객체 외부 라이브러리 의존 제거

* test: 경로 조회 예외 테스트 추가

* refactor: distance 파라미터 final

* refactor: lind Id Wrapper 클래스로 변경

* refactor: infra 패키지 분리

* refactor: 컨트롤러 테스트 응답 값 검증 추가

* refactor: 요금 계산 메서드 분리

* refactor: Path 객체 테스트 수정

+++ src/main/java/wooteco/subway/dto/PathResponse.java
@@ -9,12 +9,12 @@ public class PathResponse {
 
     private List<Station> stations;
     private int distance;
-    private int fare;
+    private double fare;
 
     public PathResponse() {
     }
 
-    public PathResponse(List<Station> stations, int distance, int fare) {
+    public PathResponse(List<Station> stations, int distance, double fare) {
         this.stations = stations;
         this.distance = distance;
         this.fare = fare;
@@ -28,7 +28,7 @@ public class PathResponse {
         return distance;
     }
 
-    public int getFare() {
+    public double getFare() {
         return fare;
     }
 }
